Fallback to default XDG_DATA_DIRS for icon search
According to the XDG Base Directory Specification, `XDG_DATA_DIRS`
should default to `/usr/local/share/:/usr/share/` if empty. I've found
that the variable is inconsistently set across environments. I couldn't
find any reason as to why, but it seems like defaulting to the above is
expected behavior.

Besides that, I switched the code that manually split the env arg with a
function in the standard library that does the same.