fix: Duplicate color wallpaper previews
Closes: #348

The issue arises after the wallpaper settings page is opened after
choosing a default color as a background. The chosen color is treated as
a custom color on reload which is then duplicated when the default
colors are displayed. Solving this issue is as simple as filtering the
customs colors list for default colors.

I unwrap two `Options` from `partial_cmp` in this fix. Floats don't
implement `Eq` due to `NaN`s but they do implement `PartialOrd` and
`PartialEq`. Colors shouldn't have `NaN`s at all, so unwrapping here is
fine as that invariant should hold unless there is a bug elsewhere.