display: Add options for DPI scales up to 300%
Scale factors above 200% may be useful on things like 4k 13 inch
laptops. So we should provide higher options, even if we don't default
to anything above 200%.

The way the dropdown is formatted with scrolling on smaller window sizes
could probably be improved. But that will presumably be a libcosmic
change, not anything here.