Reworked access to density to use page struct from the view
Instead of recomputing the whole configuration inside the view on every
render, it's possible to assign the density value when the message is
receive to the Page struct. Since the page is passed as an argument to
the View, those values are accessible to the view when the redraw
happens.

This is already how it works with Roundness and I missed it
initially, but since it's already a pattern in use in this mod, this
should make the module more consistent.