fix(applets_inner): use correct background color
This makes the applet dnd items use `bg_component_color`, to match designs and prevent them from visually merging with the context drawer when it's open.
Also makes the applet drop target a blank accent-colored container, rather than it being a copy of the item that's being dragged, making the UX feel a bit nicer.