fix(font): remove intrusive filter for fonts
Some regular fonts that have slightly higher weights get filtered out
due to the filter used in the font loading function being a little too
intrusive.

Remove the condition for font weight to be regular so that all fonts
will be available to be set for the interface/mono fonts.

Others like KDE, Gnome do not filter out fonts with higher weights since
on the same system, other DEs allow the fonts to be selected. So
there is need for Cosmic DE to filter out these fonts.

Fixes: https://github.com/pop-os/cosmic-settings/issues/859
Signed-off-by: Aurabindo Pillai <mail@aurabindo.in>