feat(shortcuts): WIP load existing shortcut into create context drawer (pop-os/cosmic-settings#1566)
An initial quick and dirty attempt to allow the descriptions and actions
of custom shortcuts to be edited after create. This needs some more work
to behave properly with the keybind replace dialog.

We hook into `ShowShortcut` and set `self.add_shortcut.active = true` to
have the correct context drawer shown. The extra code in the other
branch of `custom::Page::context_drawer` should also probably be
removed.